refactor: checkbox 리팩토링 및 변경
